+++ src/commands/resourceCommand.ts
@@ -1,6 +1,6 @@
 import * as vscode from 'vscode';
 
-import { runBash } from '../utils/utils';
+import { executeCmd } from '../utils/utils';
 import { Node } from '../models';
 import { resourceTreeView, resourceViewDataProvider } from '../view/resourceView';
 import { resourceService, configService } from '../services';
@@ -9,7 +9,7 @@ import { ListItem, NodeItem } from '../view/components';
 let autoRefreshTimer: NodeJS.Timeout;
 
 function extractNodes(nodeString: string, short_length: number, long_length: number): Node[] {
-    let slices = [short_length, short_length, short_length, short_length, long_length, long_length, short_length];
+    let slices = [short_length, short_length, short_length, short_length, long_length, long_length, short_length, short_length];
     slices.reduce((arr, currentValue, currentIndex) => {
         if (currentIndex > 0) {
             slices[currentIndex] = slices[currentIndex - 1] + currentValue;
@@ -20,14 +20,14 @@ function extractNodes(nodeString: string, short_length: number, long_length: num
 
     let nodeList: Node[] = [];
     nodeString.split('\n').forEach((value) => {
-        // JobID,Name:255,Username:20,State:20,NodeList,Gres:50,TimeLimit,TimeUsed,Command:255,STDOUT:255,STDERR:255,Reason:100
+        // NODELIST:15,Available:15,Memory:15,AllocMem:15,Gres:50,GresUsed:50,Partition:15,StateLong:15
         if (value.length === 0) { return; }
 
         let fields: string[] = slices.slice(1).reduce((arr: string[], v, i) => {
             arr.push(value.substring(v, slices[i]).trim());
             return arr;
         }, []);
-        const node = new Node(fields[0], fields[1], fields[2], fields[3], fields[4], fields[5], fields[6]);
+        const node = new Node(fields[0], fields[1], fields[2], fields[3], fields[4], fields[5], fields[6], fields[7]);
         nodeList.push(node);
     });
     return nodeList;
@@ -36,7 +36,7 @@ function extractNodes(nodeString: string, short_length: number, long_length: num
 export async function refreshResources() {
     const short = 15;
     const long = 50;
-    const [out, err] = await runBash(`sinfo --noheader --Node -O NODELIST:${short},Available:${short},Memory:${short},AllocMem:${short},Gres:${long},GresUsed:${long},Partition:${short}`);
+    const [out, err] = await executeCmd(`sinfo --noheader --Node -O NODELIST:${short},Available:${short},Memory:${short},AllocMem:${short},Gres:${long},GresUsed:${long},Partition:${short},StateLong:${short}`);
     resourceService.updateNode(...extractNodes(out, short, long));
     resourceViewDataProvider.refresh();
 }