+++ src/view/components/scriptItem.ts
@@ -3,7 +3,7 @@ import { Script } from '../../models';
 
 export class ScriptItem extends vscode.TreeItem {
     constructor(public readonly script: Script) {
-        super(script.name, vscode.TreeItemCollapsibleState.None);
+        super(script.name, vscode.TreeItemCollapsibleState.Collapsed);
         this.description = script.relativePath;
         this.contextValue = 'scriptItem';
         this.command = { command: 'slurm--.openFile', title: 'open', arguments: [script] };